Reorganize front-end JS to comply with best practices
