+++ apiConfig.html
@@ -4,6 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <title>API Configuration</title>
+  <script src="apiConfig.js"></script>
 </head>
 <body>
   <main class="noBarMain">
