+++ logs.html
@@ -1,3 +1,4 @@
 <link rel="stylesheet" href="styles/logs.css">
+<script src="logs.js"></script>
 <div id="logs">
 </div>
