+++ preloads/preload-subcontrols.js
@@ -1,6 +1,6 @@
-const { ipcRenderer } = require('electron');
+const { contextBridge, ipcRenderer } = require('electron')
 
-window.electronAPI = {
+contextBridge.exposeInMainWorld('electronAPI', {
     startTimer: () => ipcRenderer.invoke('start-timer'),
     pauseTimer: () => ipcRenderer.invoke('pause-timer'),
     clearTimer: () => ipcRenderer.invoke('clear-stored-time'),
@@ -8,4 +8,4 @@ window.electronAPI = {
     startMultiplier: (value) => ipcRenderer.send('start-multi', value),
     addTime: (amount) => ipcRenderer.send('add-time', amount),
     removeTime: (amount) => ipcRenderer.send('remove-time', amount)
-  };
+});
