+++ subathoncontrols.html
@@ -12,13 +12,13 @@
             <button onclick="skipAnimation()">Skip Animation</button>
         </div>
         <div class="column">
-            <input type="number" id="multiplier" placeholder="Multiplier (e.g., 2)" min="1">
+            <input type="number" id="multiplier" placeholder="Multiplier (e.g., 2)" required min="1">
             <button onclick="startMultiplier()">Start Multiplier</button>
             <br>
             <button onclick="stopMultiplier()">Stop Multiplier</button>
         </div>
         <div class="column">
-            <input class="hhmmss" pattern="^(\d+|((\d+:)?\d{1,2}:)?\d{1,2})$" id="timeAdjustment" placeholder="Time (HH:MM:SS, MM:SS, or seconds)">
+            <input class="hhmmss" pattern="^(\d+|((\d+:)?\d{1,2}:)?\d{1,2})$" required id="timeAdjustment" placeholder="Time (HH:MM:SS, MM:SS, or seconds)">
             <button onclick="addTime()">Add Time</button>
             <br>
             <button onclick="removeTime()">Remove Time</button>
@@ -26,50 +26,7 @@
     </div>
 
 
-    <script>
-        let multiplierInterval = null;
-
-        function startTimer() {
-            window.electronAPI.startTimer();
-        }
-
-        function pauseTimer() {
-            window.electronAPI.pauseTimer();
-        }
-
-        function clearTimer() {
-            window.electronAPI.clearTimer();
-        }
-
-        function skipAnimation() {
-            window.electronAPI.skipAnimation();
-        }
-
-       function startMultiplier() {
-            window.electronAPI.startMultiplier(document.getElementById("multiplier").value);
-       }
-
-       function stopMultiplier() {
-            window.electronAPI.startMultiplier(1);
-       }
-
-        function parseHhMmSs(v) {
-          const vs = v.split(':')
-          let secs = Number(vs.pop())
-          const minutes = vs.pop()
-          if (minutes) secs += Number(minutes) * 60
-          const hours = vs.pop()
-          if (hours) secs += Number(hours) * 60 * 60
-          return secs
-        }
-
-       function addTime() {
-            window.electronAPI.addTime(parseHhMmSs(document.getElementById("timeAdjustment").value))
-       }
-
-       function removeTime() {
-            window.electronAPI.addTime(-parseHhMmSs(document.getElementById("timeAdjustment").value))
-       }
-    </script>
+    <script src="hhmmss.js"></script>
+    <script src="subcontrols.js"></script>
 </body>
 </html>
