+++ subathonsettings.html
@@ -6,6 +6,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Subathon Settings</title>
   <link rel="stylesheet" href="global.css">
+  <script src="subsettings.js"></script>
+  <script src="hhmmss.js"></script>
 </head>
 
 <body>
