+++ timer.html
@@ -6,6 +6,7 @@
 <div id="subTimerDiv">
     <h1 id="subTimer">00:00:00</h1>
 </div>
+<script src="timer.js" encoding="utf-8"></script>
 <style>
 #warnWrap {
     text-align: right;
